Add ProjectMember.cs errors and move errors to more accurate files
